SDI-598: ✨ Upgrade to spring boot 3

+++ src/main/java/uk/gov/justice/hmpps/prison/api/model/SentenceSummary.java
@@ -93,7 +93,7 @@ public class SentenceSummary {
         private Long id;
 
         @Schema(description = "The case sequence number for the offender", example = "1")
-        private Long caseSeq;
+        private Integer caseSeq;
 
         @Schema(description = "The begin date of the court hearings", example = "2019-12-01")
         private LocalDate beginDate;
